+++ x/permission/genesis.go
@@ -18,7 +18,6 @@ func InitGenesis(ctx sdk.Context, k keeper.Keeper, genState types.GenesisState) 
 	}
 	// this line is used by starport scaffolding # genesis/module/init
 
-	// TODO consider if this is good practice, maybe using genState.AdministratorList is the correct thing to do?
 	if len(k.GetAllAdministrator(ctx)) == 0 {
 		bech32, err := sdk.AccAddressFromBech32(genState.Params.GetGenesisAdministrator())
 		if err != nil {
@@ -27,8 +26,8 @@ func InitGenesis(ctx sdk.Context, k keeper.Keeper, genState types.GenesisState) 
 
 		k.SetAdministrator(ctx, types.Administrator{
 			Address:   bech32.String(),
-			CreatedAt: int32(ctx.BlockHeight()), // TODO check if int32 instead64 is fine, less digits!
-			UpdatedAt: int32(ctx.BlockHeight()), // TODO check if int32 instead64 is fine, less digits!
+			CreatedAt: ctx.BlockHeight(),
+			UpdatedAt: ctx.BlockHeight(),
 			Blocked:   false,
 			Creator:   bech32.String(),
 		})
